afxdp/forwarding: collapse into directory module {mod,tests}.rs
Mirrors the session_glue/{mod,tests}.rs and frame/{mod,tests}.rs
patterns. forwarding.rs reaches 3,667 LOC (1,690 prod + 1,977 inline
tests + closing brace) and was the largest remaining file with a
gigantic inline #[cfg(test)] mod tests block.

Mechanical split:
- mkdir afxdp/forwarding
- afxdp/forwarding.rs -> afxdp/forwarding/mod.rs
- inline #[cfg(test)] mod tests block -> afxdp/forwarding/tests.rs (70 tests)
- afxdp.rs: #[path = "afxdp/forwarding/mod.rs"] mod forwarding
- forwarding/mod.rs ends with: #[cfg(test)] mod tests; (auto-resolved)

No production code touched. Use clauses kept verbatim at the top of
tests.rs (super::*, super::super::forwarding_build::*,
super::super::test_fixtures::*, crate::test_zone_ids::*, FabricSnapshot
etc.) — module hierarchy is unchanged because tests still resolves to
super = forwarding via mod tests.

cargo build --release: 92 warnings baseline, no errors.
cargo test --release: 865 passed, 0 failed (baseline preserved).
